sigh

+++ app/Http/Controllers/Feed/CommentController.php
@@ -10,6 +10,7 @@ use JWTAuth;
 use App;
 use App\Comment;
 use App\Commands\DeleteComment;
+use App\Http\Requests\StoreProjectRequest;
 use App\Http\Requests\StoreCommentRequest;
 use App\Http\Requests\DeleteCommentRequest;
 use App\Feed;
@@ -21,8 +22,9 @@ class CommentController extends Controller {
 	 *
 	 * @return Response
 	 */
-	public function store(Feed $feed, StoreCommentRequest $request)
+	public function store(Request $request, Feed $feed)
 	{
+		$this->validate($request, ['comment' => 'required']);
 		$user = JWTAuth::parseToken()->authenticate();
 		$comment = $this->dispatch(new PostComment($user, $request->all(), $feed));
 		return response()->json(['success' => true, 'message' => 'Comment Posted.', 'comment' => $comment]);
@@ -34,9 +36,10 @@ class CommentController extends Controller {
 	 * @param  int  $id
 	 * @return Response
 	 */
-	public function destroy(Feed $feed, Comment $comment, DeleteCommentRequest $request)
+	public function destroy(Feed $feed, Comment $comment, Request $request)
 	{
 		$user = JWTAuth::parseToken()->authenticate();
+		if($comment->owner->id != $user->id) return response()->json(['fail' => true, 'message' => 'Not authorized to delete comment.']);
 		$this->dispatch(new DeleteComment($user, $comment));
 		return response()->json(['success' => true, 'message' => 'Comment Deleted.']);
 	}
